Update impementation calculateNewFileData
This will ease the delete process, as we don't have interference from
deletions happening outside the SOURCE_FILES_DATA_GRAPH.

This is mainly tackles a potential racecondition for formData, where
there is a direct reference from the ext:SubmissionDocument to it's
source. This triple however, is managemd outside
SOURCE_FILES_DATA_GRAPH, so we cannot make all assumptions we'd like.